refactor: CouponContext 내부 로직 관심사에 따라 훅으로 분리
